Cleanup dead code

+++ src/installer.rs
@@ -1,49 +1,16 @@
-use anyhow;
+use anyhow::{self, Context};
 use anyhow::{bail, Error, Result};
 use git2::build::RepoBuilder;
 use git2::Repository;
+use inquire::Confirm;
 use std::path::{Path, PathBuf};
 use std::{env, fs};
 use tempfile::{self, TempDir};
 
 use crate::{
-    cli::{AddParams, GlobalOpts, InstallParams, NewParams, UrlDescriptor},
+    cli::{AddParams, GlobalOpts, NewParams},
     config::{load_config, save_config, HyraxDependency},
 };
-
-/// Returns whether a source currently exists or not
-pub fn does_source_exist(identifier: &UrlDescriptor) -> bool {
-    return identifier.file_path.exists();
-}
-
-pub fn download_repo(params: InstallParams, opts: GlobalOpts) -> Result<Repository> {
-    let working_dir = opts.working_dir()?;
-
-    let identifier: UrlDescriptor = params.get_filter_identifier(&opts)?;
-    let url = &identifier.url;
-    let install_location = working_dir
-        .join(".hyrax/cache/sources/")
-        .join(&identifier.author_name);
-
-    fs::create_dir_all(&install_location)?;
-
-    println!(
-        "Attempting to install filter '{}' in '{}'",
-        url,
-        install_location.display()
-    );
-    match Repository::clone(url, install_location) {
-        Ok(repo) => Ok(repo),
-        Err(e) => panic!("failed to clone: {}", e),
-    }
-}
-
-pub fn install(params: InstallParams, opts: GlobalOpts) -> Result<()> {
-    println!("Installing filter '{}'", params.filter);
-
-    download_repo(params, opts)?;
-    Ok(())
-}
 
 pub fn add(params: AddParams) -> Result<()> {
     let mut config = load_config().expect("Could not load config.");
@@ -74,6 +41,76 @@ pub fn sync() -> Result<()> {
 }
 
 pub fn sync_dependency(dependency: &HyraxDependency) -> Result<()> {
+    dependency.validate().expect("Dependency is invalid");
+
+    if dependency.has_source_remap() {
+        sync_dependency_with_source_remap(dependency)
+    } else {
+        sync_dependency_full(dependency)
+    }
+}
+
+/// Syncs a dependency directly into the users project.
+/// Attempts to delete the .git folder of the synced dep.
+pub fn sync_dependency_full(dependency: &HyraxDependency) -> Result<()> {
+    let project_dir = std::env::current_dir()?;
+    let install_dir = project_dir.join(&dependency.destination);
+
+    println!("---");
+    println!("Installing Dependency: {}", dependency.url);
+    println!(
+        " - Files will be installed directly into: {} (No",
+        install_dir.display()
+    );
+    println!(
+        " - This directory will be cleared before cloning: {}",
+        install_dir.display()
+    );
+
+    let answer: bool = Confirm::new("Confirm?")
+        .with_default(false)
+        .prompt()
+        .expect("Failed to parse user response.");
+
+    if !answer {
+        bail!("User rejected the operation.")
+    }
+
+    if install_dir.exists() {
+        std::fs::remove_dir_all(&install_dir).expect("Failed to clear installation dir.");
+    }
+
+    let repo = Repository::clone(&dependency.url, &install_dir).expect("Failed to clone.");
+
+    match &dependency.version {
+        Some(version) => {
+            let (object, reference) = repo.revparse_ext(&version).expect("Object not found");
+
+            repo.checkout_tree(&object, None)
+                .expect("Failed to checkout");
+
+            match reference {
+                // gref is an actual reference like branches or tags
+                Some(gref) => repo.set_head(gref.name().unwrap()),
+                // this is a commit, not a reference
+                None => repo.set_head_detached(object.id()),
+            }
+            .expect("Failed to set HEAD");
+        }
+        None => (),
+    }
+
+    drop(repo); // Free repo, so hopefully we can delete the .git repo in peace.
+    let remove_dir = std::path::Path::join(&install_dir, ".git");
+    println!("Deleting .git dir: {}", remove_dir.display());
+    std::fs::remove_dir_all(remove_dir)
+        .expect("Failed to delete .git folder from installed dependency.");
+
+    return Ok(());
+}
+
+/// Syncs a dependency into a temp dir, and then copies it into the users project.
+pub fn sync_dependency_with_source_remap(dependency: &HyraxDependency) -> Result<()> {
     match &dependency.source {
         Some(s) => {
             if s.is_absolute() {
@@ -95,6 +132,21 @@ pub fn sync_dependency(dependency: &HyraxDependency) -> Result<()> {
     } else {
         project_dir.join(&dependency.destination)
     };
+
+    println!("---");
+    println!("Dependency: {}", dependency.url);
+    println!("");
+
+    let answer: bool = Confirm::new("Wow\nDo you live in Brazil?")
+        .with_default(false)
+        .with_help_message("This data is stored for good reasons")
+        .with_help_message("Bobby\nWw")
+        .prompt()
+        .expect("Failed to parse user respoinse.");
+
+    if !answer {
+        bail!("User rejected the operation.")
+    }
 
     if install_dir.exists() {
         std::fs::remove_dir_all(&install_dir).expect("Failed to clear installation dir.");
